// contracts/generated/typescript/models/BlockRedirect.ts
export interface BlockRedirect {
    /**
     *
     * @type {string}
     * @memberof BlockRedirect
     */
    sourcePageId: string;
    /**
     *
     * @type {string}
     * @memberof BlockRedirect
     */
    sourceBlockId: string;
    /**
     *
     * @type {string}
     * @memberof BlockRedirect
     */
    targetPageId: string;
    /**
     *
     * @type {string}
     * @memberof BlockRedirect
     */
    targetBlockId: string;
    /**
     *
     * @type {string}
     * @memberof BlockRedirect
     */
    targetPageTitle: string;
    /**
     *
     * @type {string}
     * @memberof BlockRedirect
     */
    targetCurrentSlug: string;
    /**
     *
     * @type {string}
     * @memberof BlockRedirect
     */
    createdBy: string;
    /**
     *
     * @type {Date}
     * @memberof BlockRedirect
     */
    createdAt: Date;
}